Use metadata golang package instead of raw HTTP
